+++ pkg/resource/job.go
@@ -94,7 +94,7 @@ func (b JobBuilder) buildPodTemplate() corev1.PodTemplateSpec {
 			TerminationGracePeriodSeconds: ptr.Int64(60),
 			Containers:                    b.MakeContainers(),
 			AutomountServiceAccountToken:  ptr.Bool(false),
-			ServiceAccountName:            "cockroach-database-sa",
+			ServiceAccountName:            b.ServiceAccountName(),
 			RestartPolicy:                 corev1.RestartPolicyNever,
 		},
 	}