+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Deprecated legacy OpenShift packaging format in favor of new bundle format
 * Removed unused beta channel
+* Dynamically create instance specific service account, role, and role binding
 
 ## Fixed